fix some ui issues

+++ lib/models/personModels.dart
@@ -9,7 +9,7 @@ class personClass {
   String? gender;
   String? id;
   String? name;
-  int? score;
+  double? score;
   String? time;
   String? trackId;
   String? role;
